backend/pyproject.toml
[project]
name = "land-property-management"
version = "0.1.0"
description = "土地物业资产管理系统"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.1.0",
    "psycopg>=3.2.0",
    "asyncpg>=0.31.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "pandas>=2.0.0",
    "redis>=7.0.0",
    "pyjwt>=2.8.0",
    "aiofiles>=23.0.0",
    "psutil>=7.1.3",
    "python-dotenv>=1.2.1",
    "pillow>=12.0.0",
    "sentry-sdk[fastapi]>=1.45.0",
    "httpx>=0.28.1",
    "cryptography>=42.0.0",
    # PDF基础处理（移动到可选依赖）
    "bcrypt>=5.0.0",
]

[project.optional-dependencies]
# PDF基础处理功能（轻量级）
pdf-basic = [
    "pdfplumber>=0.11.0,<0.12.0",
    "pymupdf>=1.24.5,<1.25.0",
    "pdf2image>=1.17.0,<2.0.0",
    "markitdown>=0.1.0",
    "python-magic>=0.4.27; sys_platform != 'linux'",
    "python-magic-bin>=0.4.14; sys_platform != 'linux'",
]

# 完整功能（所有可选依赖）
all = [
    "land-property-management[pdf-basic]",
]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "pre-commit>=4.3.0",
    "bandit>=1.8.6",
    "radon>=6.0.1",
    "safety>=3.6.2",
    "testcontainers>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E402", "E501"]
per-file-ignores = {"alembic/env.py" = ["F401"]}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
# 渐进式类型检查策略：核心代码严格，边缘代码宽松
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Phase 1: 渐进式启用
check_untyped_defs = true
warn_unused_ignores = false  # Phase 2: 启用

# 启用 PEP 695 新类型参数语法 (Python 3.12+)
enable_incomplete_feature = ["NewGenericSyntax"]

# 启用 SQLAlchemy mypy 插件 (修复模型类的 valid-type 问题)
plugins = ["sqlalchemy.ext.mypy.plugin"]

# 忽略第三方库的类型错误
ignore_missing_imports = true

# === Tier 0: 核心基础设施 (STRICT) ===
# 数据库、认证、配置、模型、Schema
[[tool.mypy.overrides]]
module = [
    "src.database",
    "src.core.security",
    "src.core.config",
    "src.security.jwt_security",
    "src.models.*",
    "src.schemas.*",
]
# Phase 0: 立即启用严格模式
strict = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true

# === Tier 1: 核心业务逻辑 (STANDARD) ===
# API 端点、服务层、CRUD
[[tool.mypy.overrides]]
module = [
    "src.api.v1.*",
    "src.services.*",
    "src.crud.*",
]
# Phase 1 完成: 启用严格类型检查
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true

# === Tier 2: 边缘工具 (BASIC) ===
# Utils、Middleware、Validation
[[tool.mypy.overrides]]
module = [
    "src.utils.*",
    "src.middleware.*",
    "src.validation.*",
]
# Removed: warn_unused_configs (not allowed in per-module overrides)
# Removed: warn_redundant_casts (not allowed in per-module overrides)
# Removed: warn_unused_ignores (not allowed in per-module overrides)

[[tool.mypy.overrides]]
module = [
    "src.core.permissions",
]

# === rent_contract.py 已重构为 Mapped[] 类型 ===
# 2026-01: 完成重构，移除 misc 错误码豁免
[[tool.mypy.overrides]]
module = "src.models.rent_contract"
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "src.models.auth",
    "src.models.dynamic_permission",
    "src.core.security_event_logger",
]

[[tool.mypy.overrides]]
module = [
    "src.models.*",
]

# === 无类型第三方库的显式忽略 ===
[[tool.mypy.overrides]]
module = [
    "fitz",
    "fitz.*",
    "pymupdf",
    "pymupdf.*",
    "pdfplumber.*",
    "tenacity.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short --disable-warnings -ra --import-mode=importlib --cov=src --cov-report=term-missing --cov-report=html:../test-results/backend/coverage/html --cov-report=xml:../test-results/backend/coverage/coverage.xml --cov-report=json:../test-results/backend/coverage/coverage.json --junitxml=../test-results/backend/junit/junit-report.xml --cov-fail-under=20"
norecursedirs = ["archive", "*.egg-info", ".git", ".tox", "dist", "build", "node_modules", "__pycache__", ".pytest_cache"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
markers = [
    "unit: Unit tests - single function or class",
    "integration: Integration tests - module collaboration",
    "e2e: End-to-end tests - complete workflow",
    "slow: Slow running tests (execution time > 10s)",
    "database: Database tests - requires database connection",
    "api: API tests - API endpoint testing",
    "vision: LLM Vision API tests (Qwen, DeepSeek, GLM)",
    "pdf: PDF processing tests (markitdown, pdfplumber)",
    "rbac: RBAC tests - permission functionality",
    "security: Security tests - vulnerability and penetration testing",
    "performance: Performance tests - load and stress testing",
    "load: Load tests - high volume and sustained load testing",
    "stress: Stress tests - maximum capacity and pressure testing",
    "core: Core functionality tests",
    "asyncio: Async tests",
    "concurrency: Concurrency and stress tests"
]
minversion = "7.4"
timeout = 300
timeout_method = "thread"

[tool.coverage.run]
source = ["src"]
data_file = "../test-results/backend/coverage/.coverage"
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "*/archive/*"
]

[tool.coverage.report]
fail_under = 20
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod"
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "../test-results/backend/coverage/html"

# =============================================================================
# Bandit - Security Linter Configuration
# =============================================================================
[tool.bandit]
# Only check HIGH and MEDIUM severity issues (Phase 2: strengthened from LOW)
# - B101: assert_used (LOW severity - skip in production code)
# - B110: try_except_pass (MEDIUM severity - requires explicit action)
# - B324: hashlib.md5 (LOW severity for non-crypto use cases)
skips = ["B101"]  # Skip assert warnings in non-test code
exclude_dirs = [
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    ".mypy_cache",
    "tests",
    "migrations",
    "alembic",
]
# -ll: Only report HIGH and MEDIUM severity issues
# -iii: Skip some checks to reduce false positives
assert_used = false  # Don't treat assert as security issue

[dependency-groups]
dev = [
    "faker>=40.1.2",
]
